+++ ui/packages/shared/components/src/index.tsx
@@ -19,6 +19,7 @@ import DateTimePicker from './DateTimePicker';
 import Dropdown from './Dropdown';
 import GrpcMetadataContext, {GrpcMetadataProvider, useGrpcMetadata} from './GrpcMetadataContext';
 import Input from './Input';
+import {KeyDownProvider, useKeyDown} from './KeyDownContext';
 import Modal from './Modal';
 import ParcaContext, {defaultValue, ParcaContextProvider, useParcaContext} from './ParcaContext';
 import Pill, {PillVariant} from './Pill';
@@ -29,6 +30,7 @@ import Spinner from './Spinner';
 import Tab from './Tab';
 import Table from './Table';
 import EmptyState from './EmptyState';
+import {useURLState} from './hooks/useURLState';
 
 export type {PillVariant, SelectElement};
 
@@ -44,6 +46,7 @@ export {
   GrpcMetadataProvider,
   IconButton,
   Input,
+  KeyDownProvider,
   Modal,
   ParcaContext,
   defaultValue as parcaContextDefaultValue,
@@ -56,5 +59,7 @@ export {
   Table,
   EmptyState,
   useGrpcMetadata,
+  useKeyDown,
   useParcaContext,
+  useURLState,
 };