+++ ui/packages/shared/profile/src/GraphTooltip/index.tsx
@@ -17,8 +17,8 @@ import {usePopper} from 'react-popper';
 
 import {CallgraphNode, FlamegraphNode, FlamegraphNodeMeta, FlamegraphRootNode} from '@parca/client';
 import {getLastItem, valueFormatter} from '@parca/functions';
-import useIsShiftDown from '@parca/components/src/hooks/useIsShiftDown';
 import {hexifyAddress, truncateString} from '../';
+import {useKeyDown} from '@parca/components';
 import {
   Function as ParcaFunction,
   Location,
@@ -389,7 +389,7 @@ const GraphTooltip = ({
   );
 
   const update = popperProps.update;
-  const isShiftDown = useIsShiftDown();
+  const {isShiftDown} = useKeyDown();
 
   useEffect(() => {
     if (contextElement != null) {