+++ ui/packages/shared/profile/src/ProfileIcicleGraph/index.tsx
@@ -15,8 +15,9 @@ import {Flamegraph} from '@parca/client';
 import {useContainerDimensions} from '@parca/dynamicsize';
 
 import DiffLegend from '../components/DiffLegend';
-import IcicleGraph from '../IcicleGraph';
+import IcicleGraph from './IcicleGraph';
 import {selectQueryParam} from '@parca/functions';
+import type {NavigateFunction} from '@parca/functions';
 import {useEffect, useMemo} from 'react';
 
 const numberFormatter = new Intl.NumberFormat('en-US');
@@ -30,6 +31,8 @@ interface ProfileIcicleGraphProps {
   curPath: string[] | [];
   setNewCurPath: (path: string[]) => void;
   onContainerResize?: ResizeHandler;
+  navigateTo?: NavigateFunction;
+  loading: boolean;
 }
 
 const ProfileIcicleGraph = ({
@@ -38,6 +41,8 @@ const ProfileIcicleGraph = ({
   setNewCurPath,
   sampleUnit,
   onContainerResize,
+  navigateTo,
+  loading,
 }: ProfileIcicleGraphProps): JSX.Element => {
   const compareMode: boolean =
     selectQueryParam('compare_a') === 'true' && selectQueryParam('compare_b') === 'true';
@@ -71,8 +76,10 @@ const ProfileIcicleGraph = ({
     }, [graph]);
 
   if (graph === undefined) return <div>no data...</div>;
+
   const total = graph.total;
-  if (parseFloat(total) === 0) return <>Profile has no samples</>;
+
+  if (parseFloat(total) === 0 && !loading) return <>Profile has no samples</>;
 
   return (
     <>
@@ -89,6 +96,7 @@ const ProfileIcicleGraph = ({
           curPath={curPath}
           setCurPath={setNewCurPath}
           sampleUnit={sampleUnit}
+          navigateTo={navigateTo}
         />
       </div>
     </>